pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "m2m-migration-tool"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "renardeinside", email = "polarpersonal@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "databricks-sdk>=0.29.0",
  "python-dotenv",
  "loguru",
  "tenacity",
]

[project.urls]
Documentation = "https://github.com/renardeinside/m2m-migration-tool#readme"
Issues = "https://github.com/renardeinside/m2m-migration-tool/issues"
Source = "https://github.com/renardeinside/m2m-migration-tool"

[tool.hatch.version]
path = "src/m2m_migration_tool/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/m2m_migration_tool tests}"

[tool.coverage.run]
source_pkgs = ["m2m_migration_tool", "tests"]
branch = true
parallel = true
omit = [
  "src/m2m_migration_tool/__about__.py",
]

[tool.coverage.paths]
m2m_migration_tool = ["src/m2m_migration_tool", "*/m2m-migration-tool/src/m2m_migration_tool"]
tests = ["tests", "*/m2m-migration-tool/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
